# src/main/resources/api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Gardener API
  license:
    name: UTCN
servers:
  - url: http://gardener.swagger.io/v1
paths:
...
  /users/new:
    post:
      summary: Create a new user with the given credentials.
      operationId: signUpUser
      tags:
        - authentication
      requestBody:
        description: User credentials
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDto"
      responses:
        '200':
          description: Sign up successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/login:
    post:
      summary: Login with the given credentials. Retrieve the user token
      operationId: loginUser
      tags:
        - authentication
      requestBody:
        description: User credentials
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDto"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/logout:
    post:
      summary: Logout user with the given token
      operationId: logoutUser
      tags:
        - authentication
      parameters:
        - in: query
          name: userToken
          required: true
          schema:
            type: string
          description: The token of the user to be logged out
      responses:
        '200':
          description: Logout successful
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /plants:
    get:
      summary: List of all plant objects
      operationId: getPlants
      tags:
        - plants
      parameters:
        - in: query
          name: userToken
          required: false
          schema:
            type: string
          description: The token of the user who makes the request
      responses:
        '200':
          description: All plants of the application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantTypeDtos"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Save or update a plant object. If no id is provided, we will create a new plant. If an id is provided, we will update all the fields of the plant, this includes updating to 'null' the fields which are not sent.
      operationId: postPlant
      tags:
        - plants
      parameters:
        - in: query
          name: userToken
          required: true
          schema:
            type: string
          description: The token of the user who makes the request
      requestBody:
        description: A plant object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantTypeDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantTypeDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /plants/{id}:
    get:
      summary: One plant
      operationId: getPlant
      tags:
        - plants
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Uuid of the plant
      responses:
        '200':
          description: The plant with the matching id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantTypeDto"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete the plant with the give id
      operationId: deletePlant
      tags:
        - plants
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Uuid of the plant
        - in: query
          name: userToken
          required: true
          schema:
            type: string
          description: The token of the user who makes the request
      responses:
        '200':
          description: The plant with the matching id was deleted.
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /featureConfigurations:
    get:
      tags:
        - features
      summary: List of all feature configuration objects
      operationId: getFeatureConfigurations
      responses:
        '200':
          description: All feature configurations of the application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeaturesConfigurationDtos"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - features
      summary: Save or update a feature configuration object. If no id is provided, we will create a new feature configuration. If an id is provided, we will update all the fields of the feature configuration, this includes updating to 'null' the fields which are not sent.
      operationId: postFeatureConfiguration
      parameters:
        - in: query
          name: userToken
          required: true
          schema:
            type: string
          description: The token of the user who makes the request
      requestBody:
        description: A feature configuration object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeaturesConfigurationDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeaturesConfigurationDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /featureConfigurations/{id}:
    get:
      tags:
        - features
      summary: One feature configuration
      operationId: getFeatureConfiguration
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Uuid of the feature configuration
      responses:
        '200':
          description: The feature configuration with the matching id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeaturesConfigurationDto"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - features
      summary: Delete the feature configuration with the give id
      operationId: deleteFeatureConfiguration
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Uuid of the feature configuration
        - in: query
          name: userToken
          required: true
          schema:
            type: string
          description: The token of the user who makes the request
      responses:
        '200':
          description: The feature configuration with the matching id was deleted.
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /growingConfigurations:
    get:
      tags:
        - growingConfigs
      summary: List of all growing configurations objects
      operationId: getGrowingConfigurations
      responses:
        '200':
          description: All growing configurations of the application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrowingConfigurationDtos"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - growingConfigs
      summary: Save or update a growing configurations object. If no id is provided, we will create a new growing configuration. If an id is provided, we will update all the fields of the growing configuration, this includes updating to 'null' the fields which are not sent.
      operationId: postGrowingConfiguration
      parameters:
        - in: query
          name: userToken
          required: true
          schema:
            type: string
          description: The token of the user who makes the request
      requestBody:
        description: A growing configuration object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrowingConfigurationDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrowingConfigurationDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /growingConfigurations/{id}:
    get:
      tags:
        - growingConfigs
      summary: One growing configuration
      operationId: getGrowingConfiguration
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Uuid of the growing configuration
      responses:
        '200':
          description: The growing configuration with the matching id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrowingConfigurationDto"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - growingConfigs
      summary: Delete the growing configuration with the give id
      operationId: deleteGrowingConfiguration
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Uuid of the growing configuration
        - in: query
          name: userToken
          required: true
          schema:
            type: string
          description: The token of the user who makes the request
      responses:
        '200':
          description: The growing configuration with the matching id was deleted.
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    UserDto:
      required: ['email', 'password']
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        password:
          type: string
        token:
          type: string
    PlantTypeDto:
      required: ['name', 'featuresConfiguration', 'growingConfiguration']
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        image:
          type: string
        featuresConfiguration:
          $ref: '#/components/schemas/FeaturesConfigurationDto'
        growingConfiguration:
          $ref: '#/components/schemas/GrowingConfigurationDto'
        userId:
          type: string
          format: uuid
    PlantTypeDtos:
      type: array
      items:
        $ref: '#/components/schemas/PlantTypeDto'
    FeaturesConfigurationDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stemType:
          $ref: '#/components/schemas/StemType'
        stemSpikes:
          $ref: '#/components/schemas/StemSpikes'
        leafDivisionOfBlade:
          $ref: '#/components/schemas/LeafDivisionOfBlade'
        leafArrangementOnTheStem:
          $ref: '#/components/schemas/LeafArrangementOnTheStem'
        leafSurface:
          $ref: '#/components/schemas/LeafSurface'
        leafTexture:
          $ref: '#/components/schemas/LeafTexture'
        leafVeinType:
          $ref: '#/components/schemas/LeafVeinType'
        flowerExistance:
          $ref: '#/components/schemas/FlowerExistance'
        flowerShape:
          $ref: '#/components/schemas/FlowerShape'
        flowerPetals:
          $ref: '#/components/schemas/FlowerPetals'
        userId:
          type: string
          format: uuid
    FeaturesConfigurationDtos:
      type: array
      items:
        $ref: '#/components/schemas/FeaturesConfigurationDto'
    GrowingConfigurationDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        soilType:
          $ref: '#/components/schemas/SoilType'
        soilQuantity:
          type: integer
          format: int32
        soilHumiditySummer:
          $ref: '#/components/schemas/FiveDegreeScale'
        soilHumidityWinter:
          $ref: '#/components/schemas/FiveDegreeScale'
        airHumidity:
          $ref: '#/components/schemas/FiveDegreeScale'
        airTemperatureSummer:
          type: integer
          format: int32
        airTemperatureWinter:
          type: integer
          format: int32
        lightIntensitySummer:
          $ref: '#/components/schemas/FiveDegreeScale'
        lightIntensityWinter:
          $ref: '#/components/schemas/FiveDegreeScale'
        userId:
          type: string
          format: uuid
    GrowingConfigurationDtos:
      type: array
      items:
        $ref: '#/components/schemas/GrowingConfigurationDto'
...
    StemType:
      type: string
      enum: ['WOODY', 'HERBACEOUS', 'NOT_VISIBLE']
    StemSpikes:
      type: string
      enum: ['NO', 'WOODY', 'SOFT']
    LeafDivisionOfBlade:
      type: string
      enum: ['SIMPLE', 'WITH_LEAFLETS', 'PINNATE_COMPOUND', 'PALMATE_COMPOUND' ,'DOUBLE_COMPOUND']
    LeafArrangementOnTheStem:
      type: string
      enum: ['BASAL', 'WHORLED', 'ALTERNATE', 'OPPOSITE']
    LeafSurface:
      type: string
      enum: ['SMOOTH', 'HAIRY', 'WARTY', 'PLEATED', 'ROUGHENED_RIDGES', 'GROOVED']
    LeafTexture:
      type: string
      enum: ['LEATHERY', 'STICKY', 'FLESHY', 'GRASSY', 'NEEDLE', 'RUGOSE', 'SPIKE']
    LeafVeinType:
      type: string
      enum: ['ARCUATE', 'CROSS_VENULATE', 'DICHTOMOUS', 'LONGITUDINAL', 'PALMATE', 'PARALLEL', 'PINNATE', 'RETICULATE', 'ROTATE']
    FlowerExistance:
      type: string
      enum: ['YES', 'NO']
    FlowerShape:
      type: string
      enum: ['REGULAR', 'IREGULAR']
    FlowerPetals:
      type: string
      enum: ['UNITED', 'SEPARATED', 'FAN']
    SoilType:
      type: string
      enum: ['SUCCULENTS', 'ORCHID', 'CONIFERS', 'REGULAR']
    FiveDegreeScale:
      type: string
      enum: ['VERY_LOW', 'LOW', 'MEDIUM', 'HIGH', 'VERY_HIGH']
...
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string